backend/requirements.txt
bcrypt==4.1.2
Flask==3.0.2
psycopg2-binary==2.9.9